refactor(stress): move stress value into distinct file
